Separate ce EE saas routes
