+++ component/repo.go
@@ -1950,110 +1950,6 @@ func (c *repoComponentImpl) CreateMirror(ctx context.Context, req types.CreateMi
 	return reqMirror, nil
 }
 
-func (c *repoComponentImpl) MirrorFromSaas(ctx context.Context, namespace, name, currentUser string, repoType types.RepositoryType) error {
-	repo, err := c.repoStore.FindByPath(ctx, repoType, namespace, name)
-	if err != nil {
-		return fmt.Errorf("failed to find repo, error: %w", err)
-	}
-	m, err := c.mirrorStore.FindByRepoID(ctx, repo.ID)
-	if err != nil {
-		if !errors.Is(err, sql.ErrNoRows) {
-			return fmt.Errorf("failed to find mirror, error: %w", err)
-		}
-	}
-	if m != nil {
-		err := c.mirrorFromSaasSync(ctx, m, repo, namespace, name, repoType)
-		if err != nil {
-			return fmt.Errorf("failed to trigger mirror sync, error: %w", err)
-		}
-		return nil
-	}
-
-	var mirror database.Mirror
-	syncVersion, err := c.syncVersionStore.FindByRepoTypeAndPath(ctx, repo.PathWithOutPrefix(), repoType)
-	if err != nil {
-		return fmt.Errorf("failed to find sync version, error: %w", err)
-	}
-	mirrorSource := &database.MirrorSource{}
-	if syncVersion.SourceID == types.SyncVersionSourceOpenCSG {
-		mirrorSource.SourceName = types.OpenCSGPrefix
-	} else if syncVersion.SourceID == types.SyncVersionSourceHF {
-		mirrorSource.SourceName = types.HuggingfacePrefix
-	}
-
-	mirrorSource.SourceName = types.OpenCSGPrefix
-	syncClientSetting, err := c.syncClientSettingStore.First(ctx)
-	if err != nil {
-		return fmt.Errorf("failed to find sync client setting, error: %w", err)
-	}
-
-	sourceUrl := common.TrimPrefixCloneURLBySourceID(c.config.MultiSync.SaasSyncDomain, string(repoType), namespace, name, syncVersion.SourceID)
-	mirror.SourceUrl = sourceUrl
-	mirror.MirrorSourceID = mirrorSource.ID
-	mirror.RepositoryID = repo.ID
-	mirror.Repository = repo
-	mirror.Username = currentUser
-	mirror.AccessToken = c.config.Mirror.Token
-	mirror.SourceRepoPath = fmt.Sprintf("%s/%s", namespace, name)
-
-	taskId, err := c.git.CreateMirrorRepo(ctx, gitserver.CreateMirrorRepoReq{
-		Namespace:   namespace,
-		Name:        name,
-		CloneUrl:    mirror.SourceUrl,
-		RepoType:    repoType,
-		MirrorToken: syncClientSetting.Token,
-		Private:     false,
-	})
-	if err != nil {
-		return fmt.Errorf("failed to create mirror repo in git: %w", err)
-	}
-
-	mirror.MirrorTaskID = taskId
-
-	m, err = c.mirrorStore.Create(ctx, &mirror)
-	if err != nil {
-		return fmt.Errorf("failed to create mirror: %w", err)
-	}
-
-	mt := database.MirrorTask{
-		MirrorID: m.ID,
-		Priority: m.Priority,
-		Status:   types.MirrorQueued,
-	}
-	_, err = c.mirrorTaskStore.CancelOtherTasksAndCreate(ctx, mt)
-	if err != nil {
-		return fmt.Errorf("failed to create mirror task: %w", err)
-	}
-
-	repo.SyncStatus = types.SyncStatusPending
-
-	_, err = c.repoStore.UpdateRepo(ctx, *repo)
-	if err != nil {
-		return fmt.Errorf("failed to update repo sync status: %w", err)
-	}
-	return nil
-}
-
-func (c *repoComponentImpl) mirrorFromSaasSync(ctx context.Context, mirror *database.Mirror, repo *database.Repository, namespace, name string, repoType types.RepositoryType) error {
-	var err error
-	mt := database.MirrorTask{
-		MirrorID: mirror.ID,
-		Priority: mirror.Priority,
-		Status:   types.MirrorQueued,
-	}
-	_, err = c.mirrorTaskStore.CancelOtherTasksAndCreate(ctx, mt)
-	if err != nil {
-		return fmt.Errorf("failed to create mirror task: %w", err)
-	}
-
-	_, err = c.repoStore.UpdateRepo(ctx, *repo)
-	if err != nil {
-		return fmt.Errorf("failed to update repo sync status: %w", err)
-	}
-
-	return nil
-}
-
 func (c *repoComponentImpl) GetMirror(ctx context.Context, req types.GetMirrorReq) (*types.Mirror, error) {
 	var (
 		status      types.MirrorTaskStatus
@@ -2944,63 +2840,6 @@ func (c *repoComponentImpl) SyncMirror(ctx context.Context, repoType types.Repos
 		return fmt.Errorf("failed to create mirror task: %w", err)
 	}
 	return nil
-}
-
-func (c *repoComponentImpl) MirrorProgress(ctx context.Context, repoType types.RepositoryType, namespace, name, currentUser string) (types.LFSSyncProgressResp, error) {
-	var (
-		progressResp types.LFSSyncProgressResp
-		done         int
-	)
-	user, err := c.userStore.FindByUsername(ctx, currentUser)
-	if err != nil {
-		return progressResp, errors.New("user does not exist")
-	}
-	if !user.CanAdmin() {
-		admin, err := c.CheckCurrentUserPermission(ctx, currentUser, namespace, membership.RoleAdmin)
-		if err != nil {
-			return progressResp, fmt.Errorf("failed to check permission to create mirror, error: %w", err)
-		}
-
-		if !admin {
-			return progressResp, errorx.ErrForbiddenMsg("need be owner or admin role to check mirror progress for this repo")
-		}
-	}
-	repo, err := c.repoStore.FindByPath(ctx, repoType, namespace, name)
-	if err != nil {
-		return progressResp, fmt.Errorf("failed to find repo, error: %w", err)
-	}
-	_, err = c.mirrorStore.FindByRepoID(ctx, repo.ID)
-	if err != nil {
-		return progressResp, fmt.Errorf("failed to find mirror, error: %w", err)
-	}
-	lfsMetaObjects, err := c.lfsMetaObjectStore.FindByRepoID(ctx, repo.ID)
-	if err != nil {
-		return progressResp, fmt.Errorf("failed to find lfs meta objects by repo id, error: %w", err)
-	}
-	repoPath := fmt.Sprintf("%ss/%s", repo.RepositoryType, repo.Path)
-	for _, lfsMetaObject := range lfsMetaObjects {
-		var resp types.SingleLFSProgress
-		if lfsMetaObject.Existing {
-			resp = types.SingleLFSProgress{
-				Oid:      lfsMetaObject.Oid,
-				Progress: 100,
-				Size:     lfsMetaObject.Size,
-			}
-			done++
-		} else {
-			progress, _ := c.syncCache.GetLfsSyncFileProgress(ctx, repoPath, lfsMetaObject.Oid, strconv.Itoa(c.config.Mirror.PartSize))
-			resp = types.SingleLFSProgress{
-				Oid:      lfsMetaObject.Oid,
-				Progress: progress,
-				Size:     lfsMetaObject.Size,
-			}
-		}
-		progressResp.Progress = append(progressResp.Progress, resp)
-	}
-	progressResp.Total = len(lfsMetaObjects)
-	progressResp.Done = done
-
-	return progressResp, nil
 }
 
 func (c *repoComponentImpl) CheckDeployPermissionForUser(ctx context.Context, deployReq types.DeployActReq) (*database.User, *database.Deploy, error) {
